REFACTOR unix commands to use common pwd (#25)
* FIX opencv link path

The path for linking OpenCV was missing a folder.

* FIX command style

Some commands where not marked as such. The instructions are now more consistent.

* ADD git clone instruction

The git clone instruction ensures that people will start from a common ground and won't search for the right root folder for the following commands.

* REFACTOR linking instructions

The linking instructions change the users pwd, the pwd was not reset. Commands where refactored so that they can be copy pasted from a common root.

* REFACTOR all unix commands to use common pwd

All commands now use the same working directory, users don't have to change folder anymore and can just copy paste the commands.